Applying the initial migration

+++ API.Pagamentos/API.Pagamentos/Domain/Transaction.cs
@@ -5,9 +5,9 @@
         public long Id { get; set; }
         public double ValueTransaction { get; set; }
         public User Sender { get; set; }
-        public int SenderId { get; set; }
+        public long SenderId { get; set; }
         public User Receiver { get; set; }
-        public int ReceiverId { get; set; }
+        public long ReceiverId { get; set; }
         public DateTime LocalDateTime { get; set; }
     }
 }
